added JPetTimeWindowMC
